+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-#0.1.6
+# 0.2.0
+- complete reworked module
+- adding alive value to player
+- adding steamId to player
+- adding assist event/value
+- adding PlayerManager
+
+# 0.1.6
 - removed jsdom depency
 - removed jQuery depency